Add file checks for wget downloads in functions and dataset scripts

+++ setup-dataset.sh
@@ -28,18 +28,34 @@ echo "Getting dataset from '${base_url}'"
 
 echo "Getting traces..."
 $SUDO wget -O traces.csv "${base_url}/sampled_traces.tsv"
+if [ ! -e traces.csv ]; then
+    echo "FATAL! -> Could not get '${base_url}/sampled_traces.tsv'"
+    exit 1
+fi
 $SUDO mv traces.csv /dataset/
 
 echo "Getting deployment_ts..."
 $SUDO wget -O deployment_ts.csv "${base_url}/deployment_ts.tsv"
+if [ ! -e deployment_ts.csv ]; then
+    echo "FATAL! -> Could not get '${base_url}/deployment_ts.tsv'"
+    exit 1
+fi
 $SUDO mv deployment_ts.csv /dataset/
 
 echo "Getting min_max_normalized_service_metrics..."
 $SUDO wget -O min_max_normalized_service_metrics.csv "${base_url}/min_max_normalized_service_metrics.tsv"
+if [ ! -e min_max_normalized_service_metrics.csv ]; then
+    echo "FATAL! -> Could not get '${base_url}/min_max_normalized_service_metrics.tsv'"
+    exit 1
+fi
 $SUDO mv min_max_normalized_service_metrics.csv /dataset/
 
 echo "Getting service_graphs..."
 $SUDO wget -O service_graphs.json "${base_url}/service_graphs.json"
+if [ ! -e service_graphs.json ]; then
+    echo "FATAL! -> Could not get '${base_url}/service_graphs.json'"
+    exit 1
+fi
 $SUDO mv service_graphs.json /dataset/
 
 $SUDO ls -al /dataset/
